All the lints are fixed and all the code still runs
